+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/VerificationDetectionScanRule.java
@@ -71,16 +71,17 @@ public class VerificationDetectionScanRule extends PluginPassiveScanner {
             for (Context context : contextList) {
                 VerificationRequestDetails currentVerifDetails =
                         AuthUtils.getVerificationDetailsForContext(context.getId());
-                VerificationRequestDetails newVerfiDetails =
+                VerificationRequestDetails newVerifDetails =
                         new VerificationRequestDetails(msg, token, context);
                 if (currentVerifDetails != null
-                        && COMPARATOR.compare(newVerfiDetails, currentVerifDetails) > 0) {
+                        && newVerifDetails.getScore() > 0
+                        && COMPARATOR.compare(newVerifDetails, currentVerifDetails) > 0) {
                     // We've potentially found a better verification request
                     LOGGER.debug(
                             "Identified potentially better verification req {} for context {}",
                             msg.getRequestHeader().getURI(),
                             context.getName());
-                    AuthUtils.processVerificationDetails(context, newVerfiDetails, this);
+                    AuthUtils.processVerificationDetails(context, newVerifDetails, this);
                 }
             }
         }